+++ Toy/src/java/model/Discount.java
@@ -12,7 +12,7 @@ package model;
 public class Discount
 {
     private String discountCode;
-    private int amount;
+    private double amount;
 
     public Discount()
     {
@@ -39,12 +39,12 @@ public class Discount
         this.discountCode = discountCode;
     }
 
-    public int getAmount()
+    public double getAmount()
     {
         return amount;
     }
 
-    public void setAmount(int amount)
+    public void setAmount(double amount)
     {
         this.amount = amount;
     }
